requirements.txt
# Core - Unsloth
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git
xformers
trl
peft
accelerate
bitsandbytes

# Data & Evaluation
datasets
pandas
scikit-learn

# Config
pyyaml
